Revert bpf.c 1.84 / bpfdesc.h 1.19 for now, "panic: timeout_add: to_ticks (-1) < 0" seen by RD Thrush, http://article.gmane.org/gmane.os.openbsd.bugs/20113 where he has a long-running process using bpf which is active at the time of panic.  krw@ agrees with reverting for now.

+++ sys/net/bpfdesc.h
@@ -67,8 +67,8 @@ struct bpf_d {
 	int		bd_bufsize;	/* absolute length of buffers */
 
 	struct bpf_if *	bd_bif;		/* interface descriptor */
-	int		bd_rtout;	/* Read timeout in 'ticks' */
-	int		bd_rdStart;	/* when the read started */
+	u_long		bd_rtout;	/* Read timeout in 'ticks' */
+	u_long		bd_rdStart;	/* when the read started */
 	struct bpf_insn *bd_rfilter; 	/* read filter code */
 	struct bpf_insn *bd_wfilter; 	/* write filter code */
 	u_long		bd_rcount;	/* number of packets received */
